img_BVR/requirements.txt
pygame>=2.5.0
moderngl>=5.8.0
Pillow>=10.0.0
numpy>=1.26.0